chore: Update Docker image version to 0.1.0-rc.14 and refactor CSP middleware
This commit updates the Docker image version to "0.1.0-rc.14" in the docker-compose file and Dockerfile. It also refactors the CSP middleware, renaming `AdminCSPMiddleware` to `RelaxedCSPMiddleware` for clarity and broadening its application to other pages that require relaxed CSP policies. Additionally, the CSP configuration in production settings is updated to use `CSP_RELAXED_POLICY`, enhancing security and maintainability. New styles for the map legend are added in the CSS, improving the UI presentation.